# api/openapi.yaml
openapi: 3.0.3
info:
  title: Zenflow Devices API
  version: 1.0.0
  description: Device-profile management API.
servers:
  - url: /
security:
  - basicAuth: []
paths:
  /healthz:
    get:
      operationId: getHealth
      summary: Health check
      security: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /profiles:
    get:
      operationId: listProfiles
      summary: List profiles for the authenticated user
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createProfile
      summary: Create a new profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProfileRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
  /profiles/{id}:
    parameters:
      - $ref: '#/components/parameters/ProfileID'
    get:
      operationId: getProfile
      summary: Get a profile by ID
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    patch:
      operationId: patchProfile
      summary: Partially update a profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchProfileRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteProfile
      summary: Delete a profile
      responses:
        '204': { description: No Content }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /templates:
    get:
      operationId: listTemplates
      summary: List all templates
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateList'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /templates/{slug}:
    parameters:
      - name: slug
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getTemplate
      summary: Get a template by slug
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Offset:
      name: offset
      in: query
      required: false
      schema: { type: integer, minimum: 0, default: 0 }
    ProfileID:
      name: id
      in: path
      required: true
      schema: { type: string }
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: Internal error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    HealthResponse:
      type: object
      required: [status]
      properties:
        status: { type: string, example: ok }
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string }
        message: { type: string }
    DeviceType:
      type: string
      enum: [desktop, mobile]
    Header:
      type: object
      required: [key, value]
      properties:
        key: { type: string }
        value: { type: string }
    Profile:
      type: object
      required:
        - id
        - user_id
        - name
        - device_type
        - window_width
        - window_height
        - user_agent
        - country_code
        - custom_headers
        - created_at
        - updated_at
      properties:
        id: { type: string }
        user_id: { type: string }
        name: { type: string }
        device_type: { $ref: '#/components/schemas/DeviceType' }
        window_width: { type: integer }
        window_height: { type: integer }
        user_agent: { type: string }
        country_code: { type: string }
        custom_headers:
          type: array
          items: { $ref: '#/components/schemas/Header' }
        extra:
          type: object
          additionalProperties: true
        template_slug:
          type: string
          nullable: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    ProfileList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Profile' }
    CreateProfileRequest:
      type: object
      required:
        - name
        - device_type
        - window_width
        - window_height
        - user_agent
        - country_code
      properties:
        name: { type: string }
        device_type: { $ref: '#/components/schemas/DeviceType' }
        window_width: { type: integer }
        window_height: { type: integer }
        user_agent: { type: string }
        country_code: { type: string }
        custom_headers:
          type: array
          items: { $ref: '#/components/schemas/Header' }
        extra:
          type: object
          additionalProperties: true
        template_slug:
          type: string
          nullable: true
    PatchProfileRequest:
      type: object
      properties:
        name: { type: string }
        device_type: { $ref: '#/components/schemas/DeviceType' }
        window_width: { type: integer }
        window_height: { type: integer }
        user_agent: { type: string }
        country_code: { type: string }
        custom_headers:
          type: array
          items: { $ref: '#/components/schemas/Header' }
        extra:
          type: object
          additionalProperties: true
    Template:
      type: object
      required:
        - slug
        - name
        - device_type
        - window_width
        - window_height
        - user_agent
        - country_code
        - custom_headers
      properties:
        slug: { type: string }
        name: { type: string }
        device_type: { type: string }
        window_width: { type: integer }
        window_height: { type: integer }
        user_agent: { type: string }
        country_code: { type: string }
        custom_headers:
          type: array
          items: { $ref: '#/components/schemas/Header' }
    TemplateList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/Template' }