# internal/transport/httprest/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package gen

import (
	"context"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
)

const (
	BasicAuthScopes = "basicAuth.Scopes"
)

// Defines values for DeviceType.
const (
	Desktop DeviceType = "desktop"
	Mobile  DeviceType = "mobile"
)

// Valid indicates whether the value is a known member of the DeviceType enum.
func (e DeviceType) Valid() bool {
	switch e {
	case Desktop:
		return true
	case Mobile:
		return true
	default:
		return false
	}
}

// CreateProfileRequest defines model for CreateProfileRequest.
type CreateProfileRequest struct {
	CountryCode   string                  `json:"country_code"`
	CustomHeaders *[]Header               `json:"custom_headers,omitempty"`
	DeviceType    DeviceType              `json:"device_type"`
	Extra         *map[string]interface{} `json:"extra,omitempty"`
	Name          string                  `json:"name"`
	TemplateSlug  *string                 `json:"template_slug,omitempty"`
	UserAgent     string                  `json:"user_agent"`
	WindowHeight  int                     `json:"window_height"`
	WindowWidth   int                     `json:"window_width"`
}

// DeviceType defines model for DeviceType.
type DeviceType string

// Error defines model for Error.
type Error struct {
	Error   string `json:"error"`
	Message string `json:"message"`
}

// Header defines model for Header.
type Header struct {
	Key   string `json:"key"`
	Value string `json:"value"`
}

// HealthResponse defines model for HealthResponse.
type HealthResponse struct {
	Status string `json:"status"`
}

// PatchProfileRequest defines model for PatchProfileRequest.
type PatchProfileRequest struct {
	CountryCode   *string                 `json:"country_code,omitempty"`
	CustomHeaders *[]Header               `json:"custom_headers,omitempty"`
	DeviceType    *DeviceType             `json:"device_type,omitempty"`
	Extra         *map[string]interface{} `json:"extra,omitempty"`
	Name          *string                 `json:"name,omitempty"`
	UserAgent     *string                 `json:"user_agent,omitempty"`
	WindowHeight  *int                    `json:"window_height,omitempty"`
	WindowWidth   *int                    `json:"window_width,omitempty"`
}

// Profile defines model for Profile.
type Profile struct {
	CountryCode   string                  `json:"country_code"`
	CreatedAt     time.Time               `json:"created_at"`
	CustomHeaders []Header                `json:"custom_headers"`
	DeviceType    DeviceType              `json:"device_type"`
	Extra         *map[string]interface{} `json:"extra,omitempty"`
	Id            string                  `json:"id"`
	Name          string                  `json:"name"`
	TemplateSlug  *string                 `json:"template_slug,omitempty"`
	UpdatedAt     time.Time               `json:"updated_at"`
	UserAgent     string                  `json:"user_agent"`
	UserId        string                  `json:"user_id"`
	WindowHeight  int                     `json:"window_height"`
	WindowWidth   int                     `json:"window_width"`
}

// ProfileList defines model for ProfileList.
type ProfileList struct {
	Items []Profile `json:"items"`
}

// Template defines model for Template.
type Template struct {
	CountryCode   string   `json:"country_code"`
	CustomHeaders []Header `json:"custom_headers"`
	DeviceType    string   `json:"device_type"`
	Name          string   `json:"name"`
	Slug          string   `json:"slug"`
	UserAgent     string   `json:"user_agent"`
	WindowHeight  int      `json:"window_height"`
	WindowWidth   int      `json:"window_width"`
}

// TemplateList defines model for TemplateList.
type TemplateList struct {
	Items []Template `json:"items"`
}

// Limit defines model for Limit.
type Limit = int

// Offset defines model for Offset.
type Offset = int

// ProfileID defines model for ProfileID.
type ProfileID = string

// BadRequest defines model for BadRequest.
type BadRequest = Error

// Conflict defines model for Conflict.
type Conflict = Error

// InternalError defines model for InternalError.
type InternalError = Error

// NotFound defines model for NotFound.
type NotFound = Error

// Unauthorized defines model for Unauthorized.
type Unauthorized = Error

// ListProfilesParams defines parameters for ListProfiles.
type ListProfilesParams struct {
	Limit  *Limit  `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *Offset `form:"offset,omitempty" json:"offset,omitempty"`
}

// CreateProfileJSONRequestBody defines body for CreateProfile for application/json ContentType.
type CreateProfileJSONRequestBody = CreateProfileRequest

// PatchProfileJSONRequestBody defines body for PatchProfile for application/json ContentType.
type PatchProfileJSONRequestBody = PatchProfileRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Health check
	// (GET /healthz)
	GetHealth(w http.ResponseWriter, r *http.Request)
	// List profiles for the authenticated user
	// (GET /profiles)
	ListProfiles(w http.ResponseWriter, r *http.Request, params ListProfilesParams)
	// Create a new profile
	// (POST /profiles)
	CreateProfile(w http.ResponseWriter, r *http.Request)
	// Delete a profile
	// (DELETE /profiles/{id})
	DeleteProfile(w http.ResponseWriter, r *http.Request, id ProfileID)
	// Get a profile by ID
	// (GET /profiles/{id})
	GetProfile(w http.ResponseWriter, r *http.Request, id ProfileID)
	// Partially update a profile
	// (PATCH /profiles/{id})
	PatchProfile(w http.ResponseWriter, r *http.Request, id ProfileID)
	// List all templates
	// (GET /templates)
	ListTemplates(w http.ResponseWriter, r *http.Request)
	// Get a template by slug
	// (GET /templates/{slug})
	GetTemplate(w http.ResponseWriter, r *http.Request, slug string)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Health check
// (GET /healthz)
func (_ Unimplemented) GetHealth(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List profiles for the authenticated user
// (GET /profiles)
func (_ Unimplemented) ListProfiles(w http.ResponseWriter, r *http.Request, params ListProfilesParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Create a new profile
// (POST /profiles)
func (_ Unimplemented) CreateProfile(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Delete a profile
// (DELETE /profiles/{id})
func (_ Unimplemented) DeleteProfile(w http.ResponseWriter, r *http.Request, id ProfileID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get a profile by ID
// (GET /profiles/{id})
func (_ Unimplemented) GetProfile(w http.ResponseWriter, r *http.Request, id ProfileID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Partially update a profile
// (PATCH /profiles/{id})
func (_ Unimplemented) PatchProfile(w http.ResponseWriter, r *http.Request, id ProfileID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// List all templates
// (GET /templates)
func (_ Unimplemented) ListTemplates(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Get a template by slug
// (GET /templates/{slug})
func (_ Unimplemented) GetTemplate(w http.ResponseWriter, r *http.Request, slug string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetHealth operation middleware
func (siw *ServerInterfaceWrapper) GetHealth(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetHealth(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListProfiles operation middleware
func (siw *ServerInterfaceWrapper) ListProfiles(w http.ResponseWriter, r *http.Request) {

	var err error

	ctx := r.Context()

	ctx = context.WithValue(ctx, BasicAuthScopes, []string{})

	r = r.WithContext(ctx)

	// Parameter object where we will unmarshal all parameters from the context
	var params ListProfilesParams

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "limit", r.URL.Query(), &params.Limit, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "offset", r.URL.Query(), &params.Offset, runtime.BindQueryParameterOptions{Type: "integer", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListProfiles(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// CreateProfile operation middleware
func (siw *ServerInterfaceWrapper) CreateProfile(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BasicAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.CreateProfile(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// DeleteProfile operation middleware
func (siw *ServerInterfaceWrapper) DeleteProfile(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id ProfileID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BasicAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.DeleteProfile(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetProfile operation middleware
func (siw *ServerInterfaceWrapper) GetProfile(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id ProfileID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BasicAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetProfile(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// PatchProfile operation middleware
func (siw *ServerInterfaceWrapper) PatchProfile(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "id" -------------
	var id ProfileID

	err = runtime.BindStyledParameterWithOptions("simple", "id", chi.URLParam(r, "id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "id", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BasicAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.PatchProfile(w, r, id)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ListTemplates operation middleware
func (siw *ServerInterfaceWrapper) ListTemplates(w http.ResponseWriter, r *http.Request) {

	ctx := r.Context()

	ctx = context.WithValue(ctx, BasicAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ListTemplates(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTemplate operation middleware
func (siw *ServerInterfaceWrapper) GetTemplate(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "slug" -------------
	var slug string

	err = runtime.BindStyledParameterWithOptions("simple", "slug", chi.URLParam(r, "slug"), &slug, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "slug", Err: err})
		return
	}

	ctx := r.Context()

	ctx = context.WithValue(ctx, BasicAuthScopes, []string{})

	r = r.WithContext(ctx)

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTemplate(w, r, slug)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/healthz", wrapper.GetHealth)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/profiles", wrapper.ListProfiles)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/profiles", wrapper.CreateProfile)
	})
	r.Group(func(r chi.Router) {
		r.Delete(options.BaseURL+"/profiles/{id}", wrapper.DeleteProfile)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/profiles/{id}", wrapper.GetProfile)
	})
	r.Group(func(r chi.Router) {
		r.Patch(options.BaseURL+"/profiles/{id}", wrapper.PatchProfile)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/templates", wrapper.ListTemplates)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/templates/{slug}", wrapper.GetTemplate)
	})

	return r
}